src/requirements.txt
Werkzeug==2.2.2
Flask==2.2.2
prometheus_client==0.14.1
waitress==2.1.2
paho-mqtt==1.6.1